+++ widgetbook/lib/widgetbook/widgetbook.directories.g.dart
@@ -12,7 +12,8 @@
 import 'package:widgetbook/widgetbook.dart' as _i1;
 import 'package:widgetbook_app/widgetbook/use_cases/buttons.dart' as _i2;
 import 'package:widgetbook_app/widgetbook/use_cases/checkboxes.dart' as _i3;
-import 'package:widgetbook_app/widgetbook/use_cases/containers.dart' as _i4;
+import 'package:widgetbook_app/widgetbook/use_cases/containers.dart' as _i5;
+import 'package:widgetbook_app/widgetbook/use_cases/running_texts.dart' as _i4;
 
 final directories = <_i1.WidgetbookNode>[
   _i1.WidgetbookFolder(
@@ -50,6 +51,20 @@ final directories = <_i1.WidgetbookNode>[
           builder: _i3.checkbox,
         ),
       ),
+      _i1.WidgetbookLeafComponent(
+        name: 'NesRunningText',
+        useCase: _i1.WidgetbookUseCase(
+          name: 'default',
+          builder: _i4.normal,
+        ),
+      ),
+      _i1.WidgetbookLeafComponent(
+        name: 'NesRunningTextLines',
+        useCase: _i1.WidgetbookUseCase(
+          name: 'default',
+          builder: _i4.lines,
+        ),
+      ),
       _i1.WidgetbookFolder(
         name: 'containers',
         children: [
@@ -58,19 +73,19 @@ final directories = <_i1.WidgetbookNode>[
             useCases: [
               _i1.WidgetbookUseCase(
                 name: 'default',
-                builder: _i4.normal,
+                builder: _i5.normal,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with corner inner square painter',
-                builder: _i4.cornerInnerSquare,
+                builder: _i5.cornerInnerSquare,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with label',
-                builder: _i4.label,
+                builder: _i5.label,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with square corner painter',
-                builder: _i4.squareConer,
+                builder: _i5.squareConer,
               ),
             ],
           )